added depth limiting search
since the beginning of the game has som many possible moves it makes our agent not have enough time to look at all possible good moves making our agent not work at is best posible rank. so we are using a depth limitation for the beginning of the game and will ad a groth function the depth to make it look forward moves as the game develops.